Booking Data Science Project
Uploaded the first version of my Data Science project created with Chai Halfon. We crawled through Booking website, created graphs and visualizations, and then proceeded to try and predict the number of average available rooms and average hotel prices in each month, for the next 12 months starting from when we ran the project, using the sklearn package for the machine learning functions.